{"schema":"decodex/commit/1","summary":"Split trace compare types into owned module after strict validation.","authority":"manual"}
