+++ apps/elf-eval/src/app/types.rs
@@ -1,11 +1,8 @@
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use sqlx::FromRow;
-use time::OffsetDateTime;
 use uuid::Uuid;
 
 use crate::app::SearchMode;
-use elf_service::{RankingRequestOverride, SearchRequest, search::TraceReplayItem};
+use elf_service::{RankingRequestOverride, SearchRequest};
 
 #[derive(Debug, Deserialize)]
 pub(super) struct EvalDataset {
@@ -216,123 +213,6 @@ pub(super) struct QueryVariantDelta {
 pub(super) struct QueryStabilityDelta {
 	pub(super) positional_churn_at_k: f64,
 	pub(super) set_churn_at_k: f64,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceCompareOutput {
-	pub(super) policies: TraceComparePolicies,
-	pub(super) summary: TraceCompareSummary,
-	pub(super) traces: Vec<TraceCompareTrace>,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceComparePolicies {
-	pub(super) a: TraceComparePolicy,
-	pub(super) b: TraceComparePolicy,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceComparePolicy {
-	pub(super) config_path: String,
-	pub(super) policy_id: String,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceCompareSummary {
-	pub(super) trace_count: usize,
-	pub(super) avg_positional_churn_at_k: f64,
-	pub(super) avg_set_churn_at_k: f64,
-	pub(super) avg_a_retrieval_top3_retention: f64,
-	pub(super) avg_b_retrieval_top3_retention: f64,
-	pub(super) avg_retrieval_top3_retention_delta: f64,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceCompareTrace {
-	pub(super) trace_id: Uuid,
-	pub(super) query: String,
-	pub(super) candidate_count: u32,
-	pub(super) top_k: u32,
-	pub(super) created_at: String,
-	pub(super) a: TraceCompareVariant,
-	pub(super) b: TraceCompareVariant,
-	pub(super) churn: TraceCompareChurn,
-	pub(super) guardrails: TraceCompareGuardrails,
-	pub(super) stage_deltas: Vec<TraceCompareStageDelta>,
-	pub(super) regression_attribution: TraceCompareRegressionAttribution,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceCompareVariant {
-	pub(super) policy_id: String,
-	pub(super) items: Vec<TraceReplayItem>,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceCompareChurn {
-	pub(super) positional_churn_at_k: f64,
-	pub(super) set_churn_at_k: f64,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceCompareGuardrails {
-	pub(super) retrieval_top3_total: usize,
-	pub(super) a_retrieval_top3_retained: usize,
-	pub(super) a_retrieval_top3_retention: f64,
-	pub(super) b_retrieval_top3_retained: usize,
-	pub(super) b_retrieval_top3_retention: f64,
-	pub(super) retrieval_top3_retention_delta: f64,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceCompareStageDelta {
-	pub(super) stage_order: u32,
-	pub(super) stage_name: String,
-	pub(super) baseline_item_count: u32,
-	pub(super) a_item_count: u32,
-	pub(super) b_item_count: u32,
-	pub(super) item_count_delta: i64,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(super) baseline_stats: Option<Value>,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct TraceCompareRegressionAttribution {
-	pub(super) primary_stage: String,
-	pub(super) evidence: String,
-}
-
-#[derive(FromRow)]
-pub(super) struct TraceCompareTraceRow {
-	pub(super) trace_id: Uuid,
-	pub(super) query: String,
-	pub(super) candidate_count: i32,
-	pub(super) top_k: i32,
-	pub(super) created_at: OffsetDateTime,
-}
-
-#[derive(FromRow)]
-pub(super) struct TraceCompareCandidateRow {
-	pub(super) candidate_snapshot: Value,
-	pub(super) note_id: Uuid,
-	pub(super) chunk_id: Uuid,
-	pub(super) chunk_index: i32,
-	pub(super) snippet: String,
-	pub(super) retrieval_rank: i32,
-	pub(super) rerank_score: f32,
-	pub(super) note_scope: String,
-	pub(super) note_importance: f32,
-	pub(super) note_updated_at: OffsetDateTime,
-	pub(super) note_hit_count: i64,
-	pub(super) note_last_hit_at: Option<OffsetDateTime>,
-}
-
-#[derive(FromRow)]
-pub(super) struct TraceCompareStageRow {
-	pub(super) stage_order: i32,
-	pub(super) stage_name: String,
-	pub(super) stage_payload: Value,
-	pub(super) item_count: i64,
 }
 
 pub(super) struct MergedQuery {
